implement store name with env config
