+++ .env.example
@@ -6,3 +6,7 @@ NEXT_PUBLIC_FIREBASE_CLIENT_MESSAGING_SENDER_ID=example
 NEXT_PUBLIC_FIREBASE_CLIENT_APP_ID=example
 
 NEXT_PUBLIC_API_ENDPOINT=example
+API_ENDPOINT=example
+
+NEXT_PUBLIC_STORE_NAME=example
+STORE_NAME=example
