+++ src/app/auth/sign-in/page.tsx
@@ -3,7 +3,6 @@ import type { Metadata } from "next";
 
 import { config } from "@/config";
 import { GuestGuard } from "@/components/auth/GuestGuard";
-import { Layout } from "@/components/auth/layout";
 import { SignInForm } from "@/components/auth/SignInForm";
 
 export const metadata = {
@@ -12,10 +11,8 @@ export const metadata = {
 
 export default function Page(): React.JSX.Element {
   return (
-    <Layout>
-      <GuestGuard>
-        <SignInForm />
-      </GuestGuard>
-    </Layout>
+    <GuestGuard>
+      <SignInForm />
+    </GuestGuard>
   );
 }
