--- src/components/auth/layout.tsx
@@ -1,33 +1,0 @@
-import React from "react";
-import Box from "@mui/material/Box";
-
-export interface LayoutProps {
-  children: React.ReactNode;
-}
-
-export function Layout({ children }: LayoutProps): React.JSX.Element {
-  return (
-    <Box
-      sx={{
-        display: { xs: "flex", lg: "grid" },
-        flexDirection: "column",
-        gridTemplateColumns: "1fr 1fr",
-        minHeight: "100%",
-      }}
-    >
-      <Box sx={{ display: "flex", flex: "1 1 auto", flexDirection: "column" }}>
-        <Box
-          sx={{
-            alignItems: "center",
-            display: "flex",
-            flex: "1 1 auto",
-            justifyContent: "center",
-            p: 3,
-          }}
-        >
-          <Box sx={{ maxWidth: "450px", width: "100%" }}>{children}</Box>
-        </Box>
-      </Box>
-    </Box>
-  );
-}
