+++ src/components/navigation/config.ts
